Static analysis fixes and dependency version bumps
Automated Fixes found by running FindBugs and IntelliJ analysis tools
Upgrade dependency versions found by `mvn versions:display-dependency-updates`
Unit tests for exception flows changed to use @Rule ExpectedException so can test message
Remove deprecated JSR310 Jackson module in Jackson replaced JavaTime module

+++ izettle-java/src/main/java/com/izettle/java/Hex.java
@@ -18,8 +18,8 @@ public class Hex {
         for (int i = 0; i < byteArr.length; i++) {
             byte b1 = (byte) ((byteArr[i] >> 4) & 0x0000000F);
             byte b2 = (byte) (byteArr[i] & 0x0000000F);
-            result[2 * i] = BASE_16_DIGITS[(int) b1];
-            result[2 * i + 1] = BASE_16_DIGITS[(int) b2];
+            result[2 * i] = BASE_16_DIGITS[b1];
+            result[2 * i + 1] = BASE_16_DIGITS[b2];
         }
         return String.valueOf(result);
     }
